RPython: add tests for py2 and py3 print syntax
these need to live in separate files because from __future__ import
print_function changes the syntax
